(GH-16) Ignore data type arrays
Prior to this commit the plugin would evaluate a data types value array as a
standard array type. This would cause the line to be flagged as invalid.

When running with `--fix`, the plugin would attempt to fix this condition
by appending a comma token to the final element of the array.i

According to puppet-syntax this is invalid.

For example:

Vaid:

```
  Enum[
    'enforcing',
    'permissive',
    'disabled'
  ] $selinux_setting = 'permissive',

```

Invalid:

```
  Enum[
    'enforcing',
    'permissive',
    'disabled',
  ] $selinux_setting = 'permissive',

```

Running `bundle exec rake syntax` post fix returns the following
error:

```
Could not parse for environment *root*: Syntax error at ']' (file: ./test.pp, line: 7, column: 1)
```

The above is not true for normal array declarations. Puppet syntax will
happily parse a normal array with a trailing comma.

Valid auto-fix from the plugin

```
 $test = [
    'one',
    'two',
    'three',
  ]
```

This commit fixes the above by adding functionality to check whether the current
`array_index` is actually an array of data type values.

If the condition is true we skip the current iteration so that it will
not be linted.