(MAINT) Add Exclude setting
When inheriting from a gem, the AllCops/exclude property doesn't seem to
be propagated.

Running rubocop with --degbug confirms that it is recognising the
directive to inherit .rubocop.yml from puppet-line and rules seem to be
valid.

This commit adds the AllCops/Exclude containing a single value that
excludes the "vendor" directory that appears when running in ci.